store: Move pool::ForeignServer to its own module
No functional change
